requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
dj-database-url==2.3.0
Django==5.1.3
django-crispy-forms==2.3
django-environ==0.11.2
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==24.2
Pillow==10.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
python-decouple==3.8
python-http-client==3.3.7
sendgrid==3.6.5
sendgrid-django==4.2.0
smtpapi==0.1.2
sqlparse==0.5.1
typing_extensions==4.12.2
unittest2py3k==0.5.1
whitenoise==6.8.2